Organizaçõo/ Separação em ramos das linhas de código --- Novas telas e funcionalidades.
